uruchomiono odbiór sygnału PPM na kanale RC1. Kanał RC2 nie  działa prawdopoeobnie z powodu problemów sprzetowych. Dodano okno rysowania stanu kanałów RC

+++ CM4/Core/Src/wymiana_CM4.c
@@ -38,7 +38,7 @@ uint8_t PobierzDaneWymiany_CM7(void)
 		chErr = HAL_HSEM_Take(HSEM_CM7_TO_CM4, 0);
 		if (chErr == ERR_OK)
 		{
-			for (uint8_t n=0; n<ROZMIAR_BUF32_WYMIANY_CM7; n++)
+			for (uint16_t n=0; n<ROZMIAR_BUF32_WYMIANY_CM7; n++)
 			{
 				uDaneCM7.nSlowa[n] = nBuforWymianyCM7[n];
 			}
@@ -66,7 +66,7 @@ uint8_t UstawDaneWymiany_CM4(void)
 		chErr = HAL_HSEM_Take(HSEM_CM4_TO_CM7, 0);
 		if (chErr == ERR_OK)
 		{
-			for (uint8_t n=0; n<ROZMIAR_BUF32_WYMIANY_CM4; n++)
+			for (uint16_t n=0; n<ROZMIAR_BUF32_WYMIANY_CM4; n++)
 			{
 				nBuforWymianyCM4[n] = uDaneCM4.nSlowa[n];
 			}
